feat: moving utility functions for fetching credentials from iam into app
