--- app/__tests__/hooks/useMintBadge.test.tsx.bk
@@ -1,217 +1,0 @@
-// src/hooks/__tests__/useMintBadge.test.ts
-import { vi, describe, it, expect, Mock } from "vitest";
-import { renderHook, act } from "@testing-library/react-hooks";
-import { useMintBadge } from "../../hooks/useMintBadge";
-import { useIssueAttestation, useAttestationNonce } from "../../hooks/useIssueAttestation";
-import { jsonRequest } from "../../utils/AttestationProvider";
-import { useMessage } from "../../hooks/useMessage";
-import { useNavigateToLastStep } from "../../hooks/useNextCampaignStep";
-import { useAccount } from "wagmi";
-
-vi.mock("../../hooks/useIssueAttestation", () => ({
-  useIssueAttestation: vi.fn().mockReturnValue({
-    issueAttestation: vi.fn(),
-    needToSwitchChain: false,
-  }),
-  useAttestationNonce: vi.fn().mockReturnValue({
-    nonce: 1,
-    isLoading: false,
-    isError: false,
-    refresh: vi.fn(),
-  }),
-}));
-vi.mock("../../utils/AttestationProvider");
-vi.mock("../../hooks/useMessage");
-vi.mock("../../hooks/useNextCampaignStep");
-
-describe("useMintBadge hook", () => {
-  const mockJsonRequest = jsonRequest as Mock;
-  const mockFailure = vi.fn();
-  const mockGoToLastStep = vi.fn();
-
-  beforeEach(() => {
-    vi.resetAllMocks();
-
-    mockJsonRequest.mockImplementation(() => {});
-
-    (useMessage as Mock).mockReturnValue({
-      failure: mockFailure,
-    });
-
-    (useNavigateToLastStep as Mock).mockReturnValue(mockGoToLastStep);
-  });
-
-  it("handles successful minting", async () => {
-    // Arrange
-    const testCredentials = [
-      {
-        credential: "testCredential",
-      },
-    ];
-    const testNonce = 123;
-    const testData = { attestation: "testAttestation" };
-
-    mockGetNonce.mockResolvedValue(testNonce);
-    mockJsonRequest.mockResolvedValue({ data: testData });
-    mockIssueAttestation.mockResolvedValue(true);
-
-    const { result } = renderHook(() => useMintBadge());
-
-    // Act
-    await act(async () => {
-      await result.current.onMint({ credentials: testCredentials });
-    });
-
-    // Assert
-    expect(mockGetNonce).toHaveBeenCalled();
-    expect(mockJsonRequest).toHaveBeenCalledWith(
-      expect.any(String),
-      expect.objectContaining({
-        recipient: "0xTestAddress",
-      })
-    );
-    expect(mockIssueAttestation).toHaveBeenCalledWith({ data: testData });
-    expect(mockFailure).not.toHaveBeenCalled();
-    expect(mockGoToLastStep).toHaveBeenCalled();
-
-    // Check state updates
-    expect(result.current.syncingToChain).toBe(false);
-    expect(result.current.badgesFreshlyMinted).toBe(true);
-  });
-
-  it("shows failure message when nonce is undefined", async () => {
-    // Arrange
-    const testCredentials = [
-      {
-        credential: "testCredential",
-      },
-    ];
-
-    mockGetNonce.mockResolvedValue(undefined);
-
-    const { result } = renderHook(() => useMintBadge());
-
-    // Act
-    await act(async () => {
-      await result.current.onMint({ credentials: testCredentials });
-    });
-
-    // Assert
-    expect(mockGetNonce).toHaveBeenCalled();
-    expect(mockFailure).toHaveBeenCalledWith({
-      title: "Error",
-      message: "An unexpected error occurred while trying to get the nonce.",
-    });
-    expect(mockJsonRequest).not.toHaveBeenCalled();
-    expect(mockIssueAttestation).not.toHaveBeenCalled();
-    expect(mockGoToLastStep).not.toHaveBeenCalled();
-
-    // Check state updates
-    expect(result.current.syncingToChain).toBe(false);
-    expect(result.current.badgesFreshlyMinted).toBe(false);
-  });
-
-  it("shows failure message when attestation generation returns an error", async () => {
-    // Arrange
-    const testCredentials = [
-      {
-        credential: "testCredential",
-      },
-    ];
-    const testNonce = 123;
-
-    mockGetNonce.mockResolvedValue(testNonce);
-    mockJsonRequest.mockResolvedValue({ data: { error: "some error" } });
-
-    const { result } = renderHook(() => useMintBadge());
-
-    // Act
-    await act(async () => {
-      await result.current.onMint({ credentials: testCredentials });
-    });
-
-    // Assert
-    expect(mockGetNonce).toHaveBeenCalled();
-    expect(mockJsonRequest).toHaveBeenCalled();
-    expect(mockFailure).toHaveBeenCalledWith({
-      title: "Error",
-      message: "An unexpected error occurred while generating attestations.",
-    });
-    expect(mockIssueAttestation).not.toHaveBeenCalled();
-    expect(mockGoToLastStep).not.toHaveBeenCalled();
-
-    // Check state updates
-    expect(result.current.syncingToChain).toBe(false);
-    expect(result.current.badgesFreshlyMinted).toBe(false);
-  });
-
-  it("shows failure message when issueAttestation throws an error", async () => {
-    // Arrange
-    const testCredentials = [
-      {
-        credential: "testCredential",
-      },
-    ];
-    const testNonce = 123;
-    const testData = { attestation: "testAttestation" };
-
-    mockGetNonce.mockResolvedValue(testNonce);
-    mockJsonRequest.mockResolvedValue({ data: testData });
-    mockIssueAttestation.mockRejectedValue(new Error("Attestation failed"));
-
-    const { result } = renderHook(() => useMintBadge());
-
-    // Act
-    await act(async () => {
-      await result.current.onMint({ credentials: testCredentials });
-    });
-
-    // Assert
-    expect(mockGetNonce).toHaveBeenCalled();
-    expect(mockJsonRequest).toHaveBeenCalled();
-    expect(mockIssueAttestation).toHaveBeenCalledWith({ data: testData });
-    expect(mockFailure).toHaveBeenCalledWith({
-      title: "Error",
-      message: "An unexpected error occurred while trying to bring the data onchain.",
-    });
-    expect(mockGoToLastStep).not.toHaveBeenCalled();
-
-    // Check state updates
-    expect(result.current.syncingToChain).toBe(false);
-    expect(result.current.badgesFreshlyMinted).toBe(false);
-  });
-
-  it("handles general exception during onMint", async () => {
-    // Arrange
-    const testCredentials = [
-      {
-        credential: "testCredential",
-      },
-    ];
-    const testNonce = 123;
-
-    mockGetNonce.mockResolvedValue(testNonce);
-    mockJsonRequest.mockRejectedValue(new Error("Network error"));
-
-    const { result } = renderHook(() => useMintBadge());
-
-    // Act
-    await act(async () => {
-      await result.current.onMint({ credentials: testCredentials });
-    });
-
-    // Assert
-    expect(mockGetNonce).toHaveBeenCalled();
-    expect(mockJsonRequest).toHaveBeenCalled();
-    expect(mockIssueAttestation).not.toHaveBeenCalled();
-    expect(mockFailure).toHaveBeenCalledWith({
-      title: "Error",
-      message: "An unexpected error occurred while trying to bring the data onchain.",
-    });
-    expect(mockGoToLastStep).not.toHaveBeenCalled();
-
-    // Check state updates
-    expect(result.current.syncingToChain).toBe(false);
-    expect(result.current.badgesFreshlyMinted).toBe(false);
-  });
-});
